src/requirements.txt
absl-py==1.4.0
cffi==1.15.1
contourpy==1.2.0
customtkinter==5.2.0
cycler==0.12.1
darkdetect==0.8.0
flatbuffers==20181003210633
fonttools==4.47.2
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.25.2
opencv-python==4.5.3.56
packaging==23.2
Pillow==8.1.2
protobuf==3.20.3
pybind11==2.11.1
pycparser==2.21
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
six==1.16.0
sounddevice==0.4.6
tflite-runtime==2.12.0
tflite-support==0.4.3
zipp==3.17.0
